backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
requests
pytesseract
pillow
python-dotenv